+++ src/sync/index.ts
@@ -1,0 +1,3 @@
+export * from "./types.js";
+export * from "./remote-apply.js";
+export * from "./conflicts.js";